@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using API.DTOs;
 using API.Entities;
+using API.Helpers;
 using API.Interfaces;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
@@ -38,11 +39,39 @@ namespace API.Data
 
         }
 
-        public async Task<IEnumerable<MemberDTO>> GetMembersAsync()
+        // public async Task<IEnumerable<MemberDTO>> GetMembersAsync()
+        // {
+        //     return await _context.Users
+        //         .ProjectTo<MemberDTO>(_mapper.ConfigurationProvider)
+        //         .ToListAsync();
+        // }
+
+        public async Task<PagedList<MemberDTO>> GetMembersAsync(UserParams userParams)
         {
-            return await _context.Users
-                .ProjectTo<MemberDTO>(_mapper.ConfigurationProvider)
-                .ToListAsync();
+            // var query = _context.Users
+            //     .ProjectTo<MemberDTO>(_mapper.ConfigurationProvider)
+            //     .AsNoTracking()
+            //     .AsQueryable();
+
+            var query = _context.Users.AsQueryable();
+            
+            query = query.Where(u => u.UserName != userParams.CurrentUsername);
+            query = query.Where(u => u.Gender == userParams.Gender);
+
+            var minDob = DateTime.Today.AddYears(-userParams.MaxAge - 1);
+            var maxDob = DateTime.Today.AddYears(-userParams.MinAge);
+
+            query = query.Where(u => u.DateOfBirth >= minDob && u.DateOfBirth <= maxDob);
+
+            query = userParams.OrderBy switch
+            {
+                "created" => query.OrderByDescending(u => u.Created),
+                _ => query.OrderByDescending(u => u.LastActive)
+            };
+
+            return await PagedList<MemberDTO>.CreateAsync(query.ProjectTo<MemberDTO>(_mapper
+                .ConfigurationProvider).AsNoTracking(), 
+                    userParams.PageNumber, userParams.PageSize);
         }
 
         public async Task<AppUser> GetUserByIdAsync(int id)
