@@ -1,6 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
+import { take } from 'rxjs/operators';
 import { Member } from 'src/app/_models/member';
+import { Pagination } from 'src/app/_models/pagination';
+import { User } from 'src/app/_models/user';
+import { UserParams } from 'src/app/_models/userParams';
+import { AccountService } from 'src/app/_services/account.service';
 import { MembersService } from 'src/app/_services/members.service';
 
 @Component({
@@ -9,14 +14,29 @@ import { MembersService } from 'src/app/_services/members.service';
   styleUrls: ['./member-list.component.css']
 })
 export class MemberListComponent implements OnInit {
-  //members: Member[];
-  members$: Observable<Member[]>;
+  members: Member[];
+  //members$: Observable<Member[]>;
+  pagination: Pagination
+  // pageNumber = 1;
+  // pageSize = 5;
+  userParams: UserParams;
+  user: User;
+  genderList = [{ value: 'male', display: 'Males' }, { value: 'female', display: 'Females' }]
 
-  constructor(private memberService: MembersService) { }
+  // constructor(private memberService: MembersService, private accountService: AccountService) {
+  //   this.accountService.currentUser$.pipe(take(1)).subscribe(user => {
+  //     this.user = user;
+  //     this.userParams = new UserParams(user);
+  //   })
+  //  }
+
+  constructor(private memberService: MembersService) {
+    this.userParams = this.memberService.getUserParams();
+  }
 
   ngOnInit(): void {
-    //this.loadMembers();
-    this.members$ = this.memberService.getMembers();
+    this.loadMembers();
+    //this.members$ = this.memberService.getMembers();
   }
 
   // loadMembers() {
@@ -25,4 +45,33 @@ export class MemberListComponent implements OnInit {
   //   })
   // }
 
+  // loadMembers() {
+  //   this.memberService.getMembers(this.pageNumber, this.pageSize).subscribe(response => {
+  //     this.members = response.result;
+  //     this.pagination = response.pagination;
+  //   })
+  // }
+
+  loadMembers() {
+    this.memberService.setUserParams(this.userParams);
+    this.memberService.getMembers(this.userParams).subscribe(response => {
+      this.members = response.result;
+      this.pagination = response.pagination;
+    })
+  }
+
+  resetFilters() {
+    //this.userParams = new UserParams(this.user)
+    this.userParams = this.memberService.resetUserParams();
+    this.loadMembers();
+  }
+
+  pageChanged(event: any) {
+    //this.pageNumber = event.page;
+    this.userParams.pageNumber = event.page;
+    this.memberService.setUserParams(this.userParams);
+    this.loadMembers();
+  }
+
+  
 }
